requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
torch==2.0.1
transformers==4.30.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
datasets==2.13.1
requests==2.31.0
web3==6.7.0
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
accelerate>=0.12.0